strip .toString() from browsers
