+++ lib/server.js
@@ -1,0 +1,21 @@
+var nanoquery = require('nanoquery')
+var assert = require('assert')
+var xtend = require('xtend')
+
+var Choo = require('./browser')
+
+Choo.prototype.toString = function (location, state) {
+  this.state = xtend(this.state, state || {})
+
+  assert.notEqual(typeof window, 'object', 'choo.mount: window was found. .toString() must be called in Node, use .start() or .mount() if running in the browser')
+  assert.equal(typeof location, 'string', 'choo.toString: location should be type string')
+  assert.equal(typeof this.state, 'object', 'choo.toString: state should be type object')
+
+  this.state.href = location.replace(/\?.+$/, '')
+  this.state.query = nanoquery(location)
+  var html = this.router(location)
+  assert.ok(html, 'choo.toString: no valid value returned for the route ' + location)
+  return html.toString()
+}
+
+module.exports = Choo
